Refactor: Split up DisguiseCameraCapture.cs
