+++ DisguiseUnityRenderStream/Editor/HDRP/DisguisePostProcessEditor.cs
@@ -2,6 +2,7 @@
 using System.Collections;
 using System.Collections.Generic;
 using System.Reflection.Emit;
+using Disguise.RenderStream;
 using UnityEditor.Rendering;
 using UnityEngine;
 using UnityEngine.Rendering.HighDefinition;
